Index, Exponential, Power, Trignometric, Hyperbolic, Error, Gamma, Nearest int, Classification functions added

+++ src/libxtl/fillcopy.cpp
@@ -4,13 +4,13 @@
 #include "fillcopy.h"
 
 template<typename T>
-INLINE_FUNC void Fill_Apply(TensorRef* result, float value)
+INLINE_FUNC void Fill_Apply(TensorRef* result, double value)
 {
     auto a = xt::adapt((T*)result->buffer, result->ElementCount(), xt::no_ownership(), result->getShape());
     a.fill(value);
 }
 
-int TS_Fill(TensorRef* result, float value)
+int TS_Fill(TensorRef* result, double value)
 {
     API_BEGIN()
     SWITCH_TENSOR_TYPE_ALL_CPU(result->elementType, Fill_Apply, result, value)
